Refactor the tests of path2import
